# assets/docs/rider/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### 说明\n接口采用非标准Restful API，所有http返回代码均为`200`，当返回为非`200`时应为network错误，需要及时排查。\n\u003cbr\u003e\n接口返回说明查看 **[返回](#返回)**\n\u003cbr\u003e\n图片/附件前缀 `https://cdn.auroraride.com/`\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 接口编号\n\n第一位代表接口端分类\n\n- M 管理端\n- R 骑手端\n- E 门店端\n- C 通用\n\n第二位代表子分类（36进制）\n\n后三位代表子编号（10进制）\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n\n### 认证\n项目接口使用简单认证，认证方式为`header`中添加对应的认证`token`\n|  header   |  类型  |  接口  |\n| :-----: | :----: | :--: |\n|  X-Rider-Token   |  string   |  骑手API  |\n| X-Manager-Token | string |  后台API  |\n|  X-Employee-Token   | string |  员工API  |\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 返回\n\n一个标准的返回应包含以下结构\n\n|  字段   |  类型  |  必填  |  说明  |\n| :-----: | :----: | :--: | :--: |\n|  code   |  int   |  是  |  返回代码  |\n| message | string |  是  |  返回消息  |\n|  data   | object |  是  |  返回数据  |\n\n`code`代码取值说明\n\n| 十进制 | 十六进制 | 说明 |\n| :----: | :------: | :--: |\n| 0  |  0x000  | 请求成功 |\n| 256 |  0x100  | 请求失败 |\n| 512 |  0x200  | *需要认证(跳转登录) |\n| 768 |  0x300  | *用户被封禁 |\n| 1024 |  0x400  | 资源未获 |\n| 1280 |  0x500  | 未知错误 |\n| 1536 |  0x600  | *需要实名 |\n| 1792 |  0x700  | *需要验证 (更换设备, 需要人脸验证) |\n| 2048 |  0x800  | *需要联系人 |\n| 2304 |  0x900  | 请求过期 |\n\n\u003e 当返回值是`1792(0x700)需要人脸验证`或`1536(0x600)需要实名`的时候`data`返回值为`{\"url\": \"string\"}`, 直接跳转url\n\n\n比如：\n\u003e 默认成功返回\n```json\n{\n  \"code\": 0,\n  \"message\": \"OK\",\n  \"data\": {\n    \"status\": true\n  }\n}\n```\n",
...
    },
    "basePath": "/",
    "paths": {
...
        "/rider/v1/cabinet/fault": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cabinet - 电柜"
                ],
                "summary": "R4008 电柜故障列表",
                "operationId": "RiderCabinetFault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rider/v1/cabinet/process": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cabinet - 电柜"
                ],
                "summary": "R4002 操作换电",
                "operationId": "RiderCabinetProcess",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RiderExchangeProcessReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
        "/rider/v1/cabinet/process/status": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cabinet - 电柜"
                ],
                "summary": "R4003 换电状态",
                "operationId": "RiderCabinetProcessStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "uuid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.RiderExchangeProcessRes"
                        }
                    }
                }
            }
        },
        "/rider/v1/cabinet/process/{serial}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cabinet - 电柜"
                ],
                "summary": "R4001 获取换电信息",
                "operationId": "RiderCabinetGetProcess",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "电柜二维码",
                        "name": "serial",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.RiderExchangeInfo"
                        }
                    }
                }
            }
        },
        "/rider/v1/cabinet/report": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cabinet - 电柜"
                ],
                "summary": "R4004 电柜故障上报",
                "operationId": "CabinetReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CabinetFaultReportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/constract/{sn}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract - 合同"
                ],
                "summary": "R3004 合同签署结果",
                "operationId": "RiderContractSignResult",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "合同编号",
                        "name": "sn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/contract/sign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract - 合同"
                ],
                "summary": "R3003 签署合同",
                "operationId": "RiderContractSign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContractSignReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.ContractSignRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "adapter.CabinetBrand": {
            "type": "string",
            "enum": [
                "UNKNOWN",
                "KAIXIN",
                "YUNDONG",
                "TUOBANG",
                "XILIULOUSERV"
            ],
            "x-enum-comments": {
                "CabinetBrandXiliulouServer": "西六楼服务器版"
            },
            "x-enum-varnames": [
                "CabinetBrandUnknown",
                "CabinetBrandKaixin",
                "CabinetBrandYundong",
                "CabinetBrandTuobang",
                "CabinetBrandXiliulouServer"
            ]
        },
...
        "model.BinInfo": {
            "type": "object",
            "properties": {
                "electricity": {
                    "description": "电量",
                    "type": "number"
                },
                "index": {
                    "description": "仓位index",
                    "type": "integer"
                },
                "voltage": {
                    "description": "电压(V)",
                    "type": "number"
                }
            }
        },
...
        "model.CabinetFaultReportReq": {
            "type": "object",
            "required": [
                "cabinetId",
                "description",
                "fault"
            ],
            "properties": {
                "attachments": {
                    "description": "附件",
                    "type": "array",
                    "maxItems": 3,
                    "items": {
                        "type": "string"
                    }
                },
                "cabinetId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fault": {
                    "type": "string"
                }
            }
        },
...
        "model.ContractSignReq": {
            "type": "object",
            "required": [
                "subscribeId"
            ],
            "properties": {
                "subscribeId": {
                    "type": "integer"
                }
            }
        },
        "model.ContractSignRes": {
            "type": "object",
            "properties": {
                "effective": {
                    "description": "是否存在生效中的合同, 若返回值为true则代表无需签合同",
                    "type": "boolean"
                },
                "sn": {
                    "description": "签署识别码",
                    "type": "string"
                },
                "url": {
                    "description": "签署URL",
                    "type": "string"
                }
            }
        },
...
        "model.RiderExchangeInfo": {
            "type": "object",
            "properties": {
                "alternative": {
                    "description": "备选方案",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BinInfo"
                        }
                    ]
                },
                "batteryFullNum": {
                    "description": "总满电电池数",
                    "type": "integer"
                },
                "batteryNum": {
                    "description": "总电池数",
                    "type": "integer"
                },
                "brand": {
                    "description": "电柜型号",
                    "allOf": [
                        {
                            "$ref": "#/definitions/adapter.CabinetBrand"
                        }
                    ]
                },
                "cityId": {
                    "description": "城市ID",
                    "type": "integer"
                },
                "doors": {
                    "description": "总仓位",
                    "type": "integer"
                },
                "emptyBin": {
                    "description": "空仓位",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BinInfo"
                        }
                    ]
                },
                "full": {
                    "description": "是否有满电电池",
                    "type": "boolean"
                },
                "fullBin": {
                    "description": "满电仓位",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BinInfo"
                        }
                    ]
                },
                "health": {
                    "description": "电柜健康状态 0离线 1正常 2故障",
                    "type": "integer"
                },
                "id": {
                    "description": "电柜ID",
                    "type": "integer"
                },
                "model": {
                    "description": "电池型号",
                    "type": "string"
                },
                "name": {
                    "description": "电柜名称",
                    "type": "string"
                },
                "serial": {
                    "description": "电柜编码",
                    "type": "string"
                },
                "uuid": {
                    "description": "操作ID",
                    "type": "string"
                }
            }
        },
        "model.RiderExchangeProcessReq": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "alternative": {
                    "description": "是否使用备选方案",
                    "type": "boolean"
                },
                "uuid": {
                    "description": "操作ID",
                    "type": "string"
                }
            }
        },
        "model.RiderExchangeProcessRes": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "消息",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1:处理中 2:成功 3:失败",
                    "type": "integer"
                },
                "step": {
                    "description": "操作步骤 1:开空电仓 2:放旧电池 3:开满电仓 4:取新电池",
                    "type": "integer"
                },
                "stop": {
                    "description": "步骤是否终止",
                    "type": "boolean"
                }
            }
        },
...
        "model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}